// src/pages/main-menu/payment/components/share-to-other/models/interfaces.ts
export interface BalShareList {
  routingId?: number;
  usedAmount: string;
  prefix: string;
  dailyConsume: string;
  shareType?: "A" | "B" | "C";
  expDate: string;
  ceilLimit: number;
  ownerSubsId?: number;
  ownerPrefix?: number;
  ownerNbr?: number;
  effDate: string;
  balShareDetailId?: number;
  balTypeName: string;
  children: Child[];
  custId?: number;
  acctResDto: AcctRes;
  isCurrency: string;
  accNbr: string;
  balId?: number;
  processType: string;
  msisdn: string;
  usableLimit: number;
  balCode: string;
  cycleConsume: string;
  relaType: string;
  acctId: number;
  priority: string;
  custName: string;
  dailyCeilLimit: number;
  spId: number;
  bsTemplateId?: number;
  subsId?: number;
  acctResId?: number;
  isHybrid?: number;
  paymentForce: string;
  balShareId?: number;
  acctNbr: string;
}

export interface Child {
  routingId?: number;
  excAcctResIds: string;
  bsdWorkHour: string;
  detailEffDate: string;
  prefix: string;
  shareType?: string;
  expDate: string;
  effDate: string;
  balShareDetailId: number;
  isCurrency: string;
  loopTimeUnit: string;
  accNbr: string;
  balId: number;
  balShareOffset: string;
  balType: string;
  acctResName: string;
  loopQuantity: string;
  priority: string;
  spId: number;
  stdCode: string;
  detailExpDate: string;
  subsId: number;
  limitSubs: string;
  acctResId: number;
  imei: string;
  balShareId?: number;
}
// ...
export interface AcctRes {
  periodClass: string;
  ratioMoney: string;
  floorLimit: number;
  remindDay: string;
  maxChgValue: string;
  creditLimit: number;
  adjustType: string;
  balType: string;
  maxValue: string;
  acctResName: string;
  remindValue: string;
  overdraftFlag: string;
  priority: string;
  dailyCeilLimit: number;
  freeFlag: string;
  stdCode: string;
  acmType: string;
  resetZero: string;
  balCategory: string;
  acctResId: number;
  maxExpDate: string;
  maxAdjustValue: string;
  gracePeriod: string;
  unitTypeId: number;
  customerFlag: string;
  ratioPrecision: string;
  defaultAcctItemTypeId?: number;
  unitTypeDto: UnitType;
  ceilLimit: number;
  dailyFloorLimit: number;
  refillable: string;
  isCurrency: string;
  parentAcctResId?: number;
  isFreeUnit: string;
  storeUnit: string;
  adjustFlag: string;
  reservePecentage: string;
  usageType: string;
  clearFlag: string;
  comments: string;
  balanceAggregation: string;
  acmAmount: string;
  unitPrecision: string;
  spId: number;
  unlimitedFlag: string;
  maxRollover: string;
  acmThreshold: string;
  paymentForce: string;
  acmUnit: string;
  category: string;
  clearDays: string;
  rewardFlag: string;
  extendRule: string;
  rolloverFlag: string;
}

export interface UnitType {
  unitTypeId: number;
  comments: string;
  unitCode: string;
  unitTypeName: string;
  trafficType: string;
  stateDate: string;
  state: string;
  spId: number;
}